// front-end/packages/shared/src/models/shippingType.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 * OpenAPI documentation.
 * OpenAPI spec version: 2.0
 */

export type ShippingType = (typeof ShippingType)[keyof typeof ShippingType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ShippingType = {
  ECONOMY: "ECONOMY",
  STANDARD: "STANDARD",
  EXPRESS: "EXPRESS",
} as const;